+++ src/test/java/com/epam/reportportal/cucumber/HooksTest.java
@@ -29,6 +29,7 @@ import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
 import org.apache.commons.lang3.tuple.Pair;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -137,11 +138,15 @@ public class HooksTest {
 		TestScenarioReporter.RP.set(reportPortal);
 	}
 
+	@AfterEach
+	public void tearDown() {
+		CommonUtils.shutdownExecutorService(executorService);
+	}
+
 	@Test
 	@SuppressWarnings("unchecked")
 	public void verify_before_after_step_reported_in_steps() {
 		TestUtils.runTests(OneStepHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -208,7 +213,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_two_before_after_step_reported_in_steps() {
 		TestUtils.runTests(TwoStepHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -235,7 +239,7 @@ public class HooksTest {
 		ArgumentCaptor<StartTestItemRQ> afterStepTwoHooksCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		verify(client, times(2)).startTestItem(same(nestedSteps.get(9).getValue()), afterStepTwoHooksCaptor.capture());
 
-		verify(client, atLeast(18)).log(any(List.class));
+		verify(client, timeout(1000).atLeast(18)).log(any(List.class));
 
 		// Verify step names from the feature file
 		List<StartTestItemRQ> mainSteps = stepCaptor.getAllValues();
@@ -308,7 +312,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_two_before_after_one_before_failed_step_reported_in_steps() {
 		TestUtils.runTests(FailStepHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -429,7 +432,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_single_before_after_scenario_reported_in_steps() {
 		TestUtils.runTests(ScenarioSingleHookReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -467,7 +469,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_two_before_after_scenario_reported_in_steps() {
 		TestUtils.runTests(ScenarioTwoHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -511,7 +512,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_two_before_after_two_scenarios_reported_in_steps() {
 		TestUtils.runTests(TwoScenarioTwoHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any(StartTestItemRQ.class));
 		verify(client, times(2)).startTestItem(same(suiteId), any());
@@ -543,7 +543,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_two_before_after_one_before_failed_scenarios_reported_in_steps() {
 		TestUtils.runTests(ScenarioTwoHooksOneBeforeFailedReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -615,7 +614,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_before_after_all_reported_in_steps() {
 		TestUtils.runTests(AllHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());
@@ -629,7 +627,6 @@ public class HooksTest {
 	@SuppressWarnings("unchecked")
 	public void verify_before_after_not_reported_in_steps() {
 		TestUtils.runTests(NoHooksReporterTest.class);
-		CommonUtils.shutdownExecutorService(executorService); // Ensure everything is finished
 
 		verify(client, times(1)).startTestItem(any());
 		verify(client, times(1)).startTestItem(same(suiteId), any());