+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Fixed
 - Start time issue for Virtual parent Items, by @HardNorth
 ### Changed
-- Client version updated on [5.3.8](https://github.com/reportportal/client-java/releases/tag/5.3.8), by @HardNorth
+- Client version updated on [5.3.9](https://github.com/reportportal/client-java/releases/tag/5.3.9), by @HardNorth
 
 ## [5.3.0]
 ### Added